pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "flask-session>=0.8.0",
    "flask-sqlalchemy>=3.1.1",
    "itsdangerous>=2.2.0",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.2",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
    "werkzeug>=3.1.3",
]
